src/requirements.txt
streamlit
pillow
boto3